Change place of files
