+++ library/corrector.rb
