+++ library/teacher.rb
