--- test_books.txt
@@ -1,1 +1,0 @@
-{"books":["{\"title\":\"dasdsa\",\"author\":\"dasdsa\",\"rental\":[]}","{\"title\":\"Leviathan wakes\",\"author\":\"James S. Corey\",\"rental\":[]}"]}
